post

+++ app/Http/Controllers/TasksController.php
@@ -13,12 +13,22 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::orderBy('id', 'asc')->paginate(25);   
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', ['tasks' => $tasks, ]);
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザーを取得
+            $user = \Auth::user();
+            // ユーザーの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザーの投稿も取得するように変更しますが、現時点ではこのユーザーの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
-
     /**
      * Show the form for creating a new resource.
      */
@@ -55,14 +65,16 @@ class TasksController extends Controller
                 ->withInput();
         }
         
-        // メッセージを作成
-        $task = new Task;
-        $task->content = $request->content;
-        $task->status = $request->status;
-        $task->save();
-
+        // 認証済みユーザー（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+            'content' => $request->content,
+            'status' => $request->status
+        ]);
+        // 前のURLへリダイレクトさせる
+        return back();
+        
         // トップページへリダイレクトさせる
-        return redirect('/');
+        // return redirect('/');
     }
 
     /**
@@ -131,12 +143,18 @@ class TasksController extends Controller
      */
     public function destroy(string $id)
     {
-        // idの値でメッセージを検索して取得
+        // idの値でタスクを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        // 認証済みユーザー（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            // idの値でメッセージを検索して取得
+            // $task = Task::findOrFail($id);
+            // メッセージを削除
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
+        }
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
     }
 }
